chore(SIP): move readableNumber to util, update SipSettings styles
